Mandriva: make sure running kernel does not affect make.sh decisions
Fix make.sh to not use /proc/ when running kernel is different than the
target kernel. Also add --no-proc-probe option to allow forcing of not
using /proc even if running kernel matches.

Update fglrx.spec to use --no-proc-probe and also use --uname_a=none to
prevent make.sh from using information from "uname -a" to determine if
the kernel is SMP or not.