Ubuntu: replace separate modalias package with a package header
In Ubuntu Natty we no longer use the fglrx-modaliases package.
All modaliases are now provided through the XB-Modaliases package
header, which dh_modaliases fills in. This way the package manager,
Jockey, etc. can check whether a card is supported by fglrx without
having to install the package. It's sufficient to use APT to query
the XB-Modaliases header of fglrx. Thanks to Martin Pitt for the
patch.