Ubuntu: add support for blacklisting card ids from GRUB gfxpayload=keep
In Natty gfxpayload=keep is set by default so as to provide a framebuffer
early in the boot process. Some cards don't play nicely with it, therefore
we need to blacklist them in a file. The format is a sequence of lines of
the following form:
  vVENDORdDEVICEsvSUBVENDORsdSUBDEVICEbcBASECLASSscSUBCLASS

Blacklist lines are regex-matched against a corresponding PCI ID string.
In practice this means that you can replace any part of the ID string
with .* to match anything.

Note: The blacklist only works in Natty, where the "grub-gfxpayload-lists"
package is available. It has no effect whatsoever in previous Ubuntu
releases.